mlops_age_prediction/requirements.txt
# API 서버 관련
fastapi
uvicorn
python-multipart
pydantic

# 이미지 처리 (최신 Python 버전 환결 호환을 위해 버전 상향/유연화)
Pillow>=10.2.0

# 딥러닝 추론 (가벼운 모델 운영을 위해 CPU 패키지 권장)
transformers
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
